+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Escape all strings that use in HTML for better security (#269).
 - Fix the missing fide out transition when hide the viewer (#275).
+- Decode image name when it comes from URL (#282).
 
 ## 1.3.3 (Apr 6, 2019)